requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
mediapipe==0.10.9
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
matplotlib>=3.7.0